connected ScientistProfilePage and authenticationPage to backend
