+++ Prototype/frontend/public/sidebarIcon.png
