+++ Prototype/frontend/src/App.css
@@ -1,4 +1,4 @@
-@import "tailwindcss";
+/* @import "tailwindcss"; */
 #root {
   max-width: none;
   margin: 0 auto;
